// app-uruguay-laravel/frontend/src/interfaces/user.interfaces.ts
export interface User {
    id: number;
    uuid: string;
    name: string;
    created_at?: string;
    updated_at?: string;
    extra_attributes?: Record<string, any>;
}

export interface UserFormData extends Omit<User, 'id' | 'created_at' | 'updated_at' | 'extra_attributes'> {
    id?: number;
    photo?: File | null;
    extra_attributes?: Record<string, any>;
}

export interface UserResponse {
    data: User | User[];
    message?: string;
    success: boolean;
}